backend/requirements.txt
# --- Core server ---
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.2.1
python-socketio==5.11.4
pydantic==2.12.3
redis==5.0.1
requests==2.32.5
tqdm==4.67.1

# --- Audio / ML Utils ---
ffmpeg-python==0.2.0
soundfile==0.13.1

# --- Numpy / SciPy required by whisperx ---
numpy==2.0.2
scipy==1.13.1

# --- WhisperX (CPU) ---
whisperx==3.7.4

# --- Demucs (optional heavy) - keep but may require extra steps; see extras file below ---
demucs==4.0.0

# --- TensorFlow CPU (if you need TF) ---
tensorflow-cpu==2.9.3
tensorflow-io-gcs-filesystem==0.31.0

# --- Database + Utils ---
mysql-connector-python==8.3.0
PyYAML==6.0.3
coloredlogs==15.0.1
colorama==0.4.6
humanfriendly==10.0

# --- Google APIs ---
google-api-python-client==2.88.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
